+++ src/components/common/Navbar.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import styled from "styled-components";
 import { Link } from "react-router-dom";
-import StarHubIconNavbar from "../assets/icons/StarHubIconNavbar.png";
-import { User } from '../types';
-import { useLogout } from '../hooks/api/useUser';
+import StarHubIconNavbar from "../../assets/icons/StarHubIconNavbar.png";
+import { User } from '../../types';
+import { useLogout } from '../../hooks/api/useUser';
 import { useQueryClient } from '@tanstack/react-query';
 
 const NavContainer = styled.nav`
