+++ src/components/main/InformCard.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import styled from "styled-components";
-import StackIcon from "../assets/icons/StackIcon.png";
-import FinishIcon from "../assets/icons/FinishIcon.png";
-import PlaceIcon from "../assets/icons/PlaceIcon.png";
-import PeopleIcon from "../assets/icons/PeopleIcon.png";
-import DuringIcon from "../assets/icons/DuringIcon.png";
+import StackIcon from "../../assets/icons/StackIcon.png";
+import FinishIcon from "../../assets/icons/FinishIcon.png";
+import PlaceIcon from "../../assets/icons/PlaceIcon.png";
+import PeopleIcon from "../../assets/icons/PeopleIcon.png";
+import DuringIcon from "../../assets/icons/DuringIcon.png";
 import { useNavigate } from "react-router-dom";
-import { InformCardProps, IconType, IconStyle } from "../types";
+import { InformCardProps, IconType, IconStyle } from "../../types";
 
 const PageContainer = styled.div`
   width: 335px;
