+++ src/components/main/OverCard.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import styled from "styled-components";
-import StackIcon from "../assets/icons/StackIcon.png";
-import FinishIcon from "../assets/icons/FinishIcon.png";
-import PlaceIcon from "../assets/icons/PlaceIcon.png";
-import PeopleIcon from "../assets/icons/PeopleIcon.png";
-import DuringIcon from "../assets/icons/DuringIcon.png";
+import StackIcon from "../../assets/icons/StackIcon.png";
+import FinishIcon from "../../assets/icons/FinishIcon.png";
+import PlaceIcon from "../../assets/icons/PlaceIcon.png";
+import PeopleIcon from "../../assets/icons/PeopleIcon.png";
+import DuringIcon from "../../assets/icons/DuringIcon.png";
 // import { Link } from "react-router-dom";
 import { useNavigate } from "react-router-dom";
